- the parameter that affects the operation of the algorithm has been increased - audio file processing has been reduced, now half of the track is processed - Many tests were written to check the algorithm's performance - sometimes the algorithm produces unexpected results from my point of view, but everything will still be tested and improved.
All of the above applies to the very first version of the algorithm 1.0.0, will be improved later

Also wrote an automated track loader to speed up development, while in version 1.0.0, it allows downloading from one pirate site and even then, with some manual code changes, but later it will be done automatically with this site, and even later, possibly from any site where it is possible to install audio. All this is done for the sake of ease of testing.